修复点击脱岗监测闪退问题，添加报警详情页面点击查看放大图功能

+++ view/client/src/main/java/net/eanfang/client/ui/activity/leave_post/LeavePostApi.java
@@ -114,7 +114,7 @@ public interface LeavePostApi {
      * @return
      */
     @POST("/yaf_station/stationDetectAlerts/contactsList")
-    Observable<BaseResponseBody<LeavePostDeviceInfoBean>> contactsList(@Query("alertId") Long alertId);
+    Observable<BaseResponseBody<LeavePostDeviceInfoBean>> contactsList(@Query("alertId") int alertId);
 
     /**
      * 首页最新脱岗警报
